docs: more javadoc, moved tasklistUrlResolver to Autoconfigured bean
