+++ engine/taskpool-collector/src/main/kotlin/io/holunda/camunda/taskpool/TaskCollectorConfiguration.kt
@@ -30,9 +30,15 @@ open class TaskCollectorConfiguration(
 
   private val logger: Logger = LoggerFactory.getLogger(TaskCollectorConfiguration::class.java)
 
+  /**
+   * Create accumulator.
+   */
   @Bean
   open fun commandAccumulator(): CommandAccumulator = ProjectingCommandAccumulator()
 
+  /**
+   * Create enricher.
+   */
   @Bean
   @ConditionalOnExpression("'\${camunda.taskpool.collector.enricher.type}' != 'custom'")
   open fun processVariablesEnricher(): VariablesEnricher =
@@ -42,6 +48,9 @@ open class TaskCollectorConfiguration(
       else -> throw IllegalStateException("Could not initialize enricher, used ${properties.enricher.type} type.")
     }
 
+  /**
+   * Create command sender.
+   */
   @Bean
   @ConditionalOnExpression("'\${camunda.taskpool.collector.sender.type}' != 'custom'")
   open fun txCommandSender(commandListGateway: CommandListGateway, accumulator: CommandAccumulator): CommandSender =
@@ -50,18 +59,6 @@ open class TaskCollectorConfiguration(
       else -> throw IllegalStateException("Could not initialize sender, used ${properties.sender.type} type.")
     }
 
-
-  @Bean
-  @ConditionalOnMissingBean
-  open fun propertyBasedTasklistUrlResolver(): TasklistUrlResolver {
-    return if (properties.tasklistUrl == null) {
-      throw IllegalStateException("Either set camunda.taskpool.collector.tasklist-url property or provide own implementation of TasklistUrlResolver")
-    } else {
-      object : TasklistUrlResolver {
-        override fun getTasklistUrl(): String = properties.tasklistUrl!!
-      }
-    }
-  }
 
   @PostConstruct
   open fun printSenderConfiguration() {
