+++ engine/taskpool-collector/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,4 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
   io.holunda.camunda.taskpool.FallbackProcessVariablesCorrelatorConfiguration,\
-  io.holunda.camunda.taskpool.FallbackProcessVariablesFilterConfiguration
+  io.holunda.camunda.taskpool.FallbackProcessVariablesFilterConfiguration, \
+  io.holunda.camunda.taskpool.FallbackTasklistUrlResolverConfiguration
